requirements_dev.txt
# Development dependencies
ruff>=0.1.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-homeassistant-custom-component>=0.13.0

# Runtime dependencies (for type checking and testing)
paho-mqtt>=2.0.0
pyhamtools>=0.11.0
websockets

# Home Assistant (for HACS component testing)
homeassistant>=2024.1.0
